TIDIJ-199 Changed structure of test folder.
